+++ src/event-engine/states/receive_stopped.ts
@@ -4,23 +4,11 @@
  * @internal
  */
 
-import { State } from '../core/state';
-import { Effects } from '../effects';
-import { Events, reconnect, restore, subscriptionChange, unsubscribeAll } from '../events';
-import { HandshakingState } from './handshaking';
-import { UnsubscribedState } from './unsubscribed';
-import * as Subscription from '../../core/types/api/subscription';
+import { reconnect, restore, subscriptionChange, unsubscribeAll } from '../events';
+import { HandshakingState, ReceiveStoppedState, UnsubscribedState } from './instances';
 
-/**
- * Context which represent current Subscription Event Engine data state.
- *
- * @internal
- */
-export type ReceiveStoppedStateContext = {
-  channels: string[];
-  groups: string[];
-  cursor: Subscription.SubscriptionCursor;
-};
+export { ReceiveStoppedState };
+export type { ReceiveStoppedStateContext } from './instances';
 
 /**
  * Stopped real-time updates (disconnected) state.
@@ -30,7 +18,6 @@ export type ReceiveStoppedStateContext = {
  *
  * @internal
  */
-export const ReceiveStoppedState = new State<ReceiveStoppedStateContext, Events, Effects>('RECEIVE_STOPPED');
 
 ReceiveStoppedState.on(subscriptionChange.type, (context, { payload }) => {
   if (payload.channels.length === 0 && payload.groups.length === 0) return UnsubscribedState.with(undefined);